feat: new `FiniteField` trait
Now everything compiles again. Will work to clean this all up and get all the tests to pass.
